Moved ClassLoading and Reflections Helper functions to common (#7103)
* Moved ClassLoading and Reflections Helper functions to common

* Fix tests

* Fix test
