+++ pulsar-functions/utils/src/main/java/org/apache/pulsar/functions/utils/ValidatorUtils.java
@@ -23,6 +23,8 @@ import lombok.extern.slf4j.Slf4j;
 import net.jodah.typetools.TypeResolver;
 import org.apache.pulsar.client.api.Schema;
 import org.apache.pulsar.common.schema.SchemaType;
+import org.apache.pulsar.common.util.ClassLoaderUtils;
+import org.apache.pulsar.common.util.Reflections;
 import org.apache.pulsar.functions.api.SerDe;
 import org.apache.pulsar.functions.proto.Function;
 import org.apache.pulsar.io.core.Sink;
@@ -31,7 +33,7 @@ import org.apache.pulsar.io.core.Source;
 import static org.apache.commons.lang3.StringUtils.isBlank;
 import static org.apache.commons.lang3.StringUtils.isEmpty;
 import static org.apache.commons.lang3.StringUtils.isNotBlank;
-import static org.apache.pulsar.functions.utils.Reflections.createInstance;
+import static org.apache.pulsar.common.util.Reflections.createInstance;
 
 @Slf4j
 public class ValidatorUtils {
@@ -43,7 +45,7 @@ public class ValidatorUtils {
             // If it's empty, we use the default schema and no need to validate
             // If it's built-in, no need to validate
         } else {
-            FunctionCommon.implementsClass(schemaType, Schema.class, clsLoader);
+            ClassLoaderUtils.implementsClass(schemaType, Schema.class, clsLoader);
             validateSchemaType(schemaType, typeArg, clsLoader, input);
         }
     }
@@ -62,13 +64,13 @@ public class ValidatorUtils {
         if (isEmpty(inputSerializer)) return;
         if (inputSerializer.equals(DEFAULT_SERDE)) return;
         try {
-            Class<?> serdeClass = FunctionCommon.loadClass(inputSerializer, clsLoader);
+            Class<?> serdeClass = ClassLoaderUtils.loadClass(inputSerializer, clsLoader);
         } catch (ClassNotFoundException | NoClassDefFoundError e) {
             throw new IllegalArgumentException(
                     String.format("The input serialization/deserialization class %s does not exist",
                             inputSerializer));
         }
-        FunctionCommon.implementsClass(inputSerializer, SerDe.class, clsLoader);
+        ClassLoaderUtils.implementsClass(inputSerializer, SerDe.class, clsLoader);
 
         SerDe serDe = (SerDe) Reflections.createInstance(inputSerializer, clsLoader);
         if (serDe == null) {
